fix(cli): gate `raven check` R init on packages.enabled
Honor packages.enabled in maybe_init_r so a user who disabled package
awareness in their editor doesn't get package/base-symbol diagnostics in
CI, matching backend::rebuild_package_library. Returns the default empty
library (not-ready) before any R discovery when disabled. Adds the
maybe_init_r_skips_when_packages_disabled regression test and documents
the behavior in docs/cli.md.

Also relocates the should_skip_directory test from cli::analysis_stats to
state (next to the function it covers).

Preserves the prior session's working tree ahead of the
build_package_library helper extraction (see
docs/superpowers/specs/2026-05-29-package-library-helper-design.md).
